refactor(format): rename `AgeKey` to `AgeFile`
